Merge freshmeat support patch by Pascal Bleser.
It is the original patch with one small change.
As fm:// can't retrieve the package list (would be too big anyways)
we still implement packages() but throw (like not supported).

The table builder now when checking if that source has the package we
need to query the package, just tries to gather the version if checking
that the package is there is not supported.

That way we keep the semantics of package(). If fm:// is used as $1
(base package list) then packages() will throw, because the plugin only
calls packages() when resolving those macros, which is fine, as fm://
does not support the macro, so you need to specify the packages manually
which does not call packages().